Now uses doubles.

+++ Source/FastNoisePlugin/Private/FastNoiseGraphFunctionLibrary.cpp
@@ -7,36 +7,36 @@
 
 FLambdaOperation2D::FLambdaOperation2D()
 {
-	OperationDelegate.BindLambda([=](const float X, const float Y)
+	OperationDelegate.BindLambda([=](const double X, const double Y)
 	{
-		return 0.f;
+		return 0.0;
 	});
 }
 
-FLambdaOperation2D::FLambdaOperation2D(float DefaultValue)
+FLambdaOperation2D::FLambdaOperation2D(double DefaultValue)
 {
-	OperationDelegate.BindLambda([=](const float X, const float Y)
+	OperationDelegate.BindLambda([=](const double X, const double Y)
 	{
 		return DefaultValue;
 	});
 }
 
-float FLambdaOperation2D::GetValue(const float X, const float Y) const
+double FLambdaOperation2D::GetValue(const double X, const double Y) const
 {
 	if (OperationDelegate.IsBound())
 	{
 		return OperationDelegate.Execute(X, Y);
 	}
 
-	return 0.f;
+	return 0.0;
 }
 
-float FLambdaOperation2D::GetValue(const FVector2D& Vector2D) const
+double FLambdaOperation2D::GetValue(const FVector2D& Vector2D) const
 {
 	return GetValue(Vector2D.X, Vector2D.Y);
 }
 
-FLambdaOperation2D UFastNoiseGraphFunctionLibrary::MakeLambdaOperation2D(float Value)
+FLambdaOperation2D UFastNoiseGraphFunctionLibrary::MakeLambdaOperation2D(double Value)
 {
 	return FLambdaOperation2D(Value);
 }
@@ -44,9 +44,9 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::MakeLambdaOperation2D(float V
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::ToScalar_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
-		return (NoiseOperation2D.GetValue(X, Y) + 1.f) / 2.f;
+		return (NoiseOperation2D.GetValue(X, Y) + 1.0) / 2.0;
 	});
 	return Operation;
 }
@@ -54,7 +54,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::ToScalar_FLambdaOperation2D(F
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Sin_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Sin(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -64,7 +64,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Sin_FLambdaOperation2D(FLambd
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Asin_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Asin(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -74,7 +74,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Asin_FLambdaOperation2D(FLamb
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Sinh_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Sinh(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -84,7 +84,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Sinh_FLambdaOperation2D(FLamb
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Cos_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Cos(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -94,7 +94,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Cos_FLambdaOperation2D(FLambd
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Acos_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Acos(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -104,7 +104,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Acos_FLambdaOperation2D(FLamb
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Tan_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Tan(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -114,7 +114,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Tan_FLambdaOperation2D(FLambd
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Atan_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Atan(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -124,7 +124,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Atan_FLambdaOperation2D(FLamb
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Atan2_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Atan2(A.GetValue(X, Y), B.GetValue(X, Y));
 	});
@@ -134,7 +134,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Atan2_FLambdaOperation2DFLamb
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::SquareRoot_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Sqrt(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -144,7 +144,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::SquareRoot_FLambdaOperation2D
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::InverseSquareRoot_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::InvSqrt(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -154,7 +154,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::InverseSquareRoot_FLambdaOper
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::MultiplyMultiply_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D Base, FLambdaOperation2D Exponent)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Pow(Base.GetValue(X, Y), Exponent.GetValue(X, Y));
 	});
@@ -164,7 +164,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::MultiplyMultiply_FLambdaOpera
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Floor_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Floor(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -174,10 +174,10 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Floor_FLambdaOperation2D(FLam
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Percent_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
-		const float ValueB = B.GetValue(X, Y);
-		return (ValueB != 0.f) ? FMath::Fmod(A.GetValue(X, Y), ValueB) : 0.f;
+		const double ValueB = B.GetValue(X, Y);
+		return (ValueB != 0.0) ? FMath::Fmod(A.GetValue(X, Y), ValueB) : 0.0;
 	});
 	return Operation;
 }
@@ -185,7 +185,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Percent_FLambdaOperation2DFLa
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Fraction_FLambdaOperation2D(FLambdaOperation2D NoiseOperation2D)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Fractional(NoiseOperation2D.GetValue(X, Y));
 	});
@@ -195,7 +195,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Fraction_FLambdaOperation2D(F
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Add_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) + B.GetValue(X, Y);
 	});
@@ -205,7 +205,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Add_FLambdaOperation2DFLambda
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Subtract_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) - B.GetValue(X, Y);
 	});
@@ -215,7 +215,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Subtract_FLambdaOperation2DFL
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Multiply_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) * B.GetValue(X, Y);
 	});
@@ -225,7 +225,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Multiply_FLambdaOperation2DFL
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Divide_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) / B.GetValue(X, Y);
 	});
@@ -235,7 +235,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Divide_FLambdaOperation2DFLam
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::SafeDivide_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return UKismetMathLibrary::SafeDivide(A.GetValue(X, Y), B.GetValue(X, Y));
 	});
@@ -245,7 +245,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::SafeDivide_FLambdaOperation2D
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Less_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B, FLambdaOperation2D TrueOperation, FLambdaOperation2D FalseOperation)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) < B.GetValue(X, Y) ? TrueOperation.GetValue(X, Y) : FalseOperation.GetValue(X, Y);
 	});
@@ -255,7 +255,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Less_FLambdaOperation2DFLambd
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::LessEqual_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B, FLambdaOperation2D TrueOperation, FLambdaOperation2D FalseOperation)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) <= B.GetValue(X, Y) ? TrueOperation.GetValue(X, Y) : FalseOperation.GetValue(X, Y);
 	});
@@ -265,7 +265,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::LessEqual_FLambdaOperation2DF
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Greater_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B, FLambdaOperation2D TrueOperation, FLambdaOperation2D FalseOperation)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) > B.GetValue(X, Y) ? TrueOperation.GetValue(X, Y) : FalseOperation.GetValue(X, Y);
 	});
@@ -275,7 +275,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Greater_FLambdaOperation2DFLa
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::GreaterEqual_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B, FLambdaOperation2D TrueOperation, FLambdaOperation2D FalseOperation)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) >= B.GetValue(X, Y) ? TrueOperation.GetValue(X, Y) : FalseOperation.GetValue(X, Y);
 	});
@@ -285,7 +285,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::GreaterEqual_FLambdaOperation
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::EqualEqual_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B, FLambdaOperation2D TrueOperation, FLambdaOperation2D FalseOperation)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) == B.GetValue(X, Y) ? TrueOperation.GetValue(X, Y) : FalseOperation.GetValue(X, Y);
 	});
@@ -295,7 +295,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::EqualEqual_FLambdaOperation2D
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::NotEqual_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B, FLambdaOperation2D TrueOperation, FLambdaOperation2D FalseOperation)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return A.GetValue(X, Y) != B.GetValue(X, Y) ? TrueOperation.GetValue(X, Y) : FalseOperation.GetValue(X, Y);
 	});
@@ -305,9 +305,9 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::NotEqual_FLambdaOperation2DFL
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::InRange_FLambdaOperation2DFLambdaOperation2D(FLambdaOperation2D In, FLambdaOperation2D Min, FLambdaOperation2D Max, FLambdaOperation2D TrueOperation, FLambdaOperation2D FalseOperation)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
-		const float InValue = In.GetValue(X, Y);
+		const double InValue = In.GetValue(X, Y);
 		return (InValue > Min.GetValue(X, Y) && InValue < Max.GetValue(X, Y)) ? TrueOperation.GetValue(X, Y) : FalseOperation.GetValue(X, Y);
 	});
 	return Operation;
@@ -316,7 +316,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::InRange_FLambdaOperation2DFLa
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Min_FLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Min(A.GetValue(X, Y), B.GetValue(X, Y));
 	});
@@ -326,7 +326,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Min_FLambdaOperation2D(FLambd
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Max_FLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Max(A.GetValue(X, Y), B.GetValue(X, Y));
 	});
@@ -336,7 +336,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Max_FLambdaOperation2D(FLambd
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Clamp_FLambdaOperation2D(FLambdaOperation2D Value, FLambdaOperation2D Min, FLambdaOperation2D Max)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Clamp(Value.GetValue(X, Y), Min.GetValue(X, Y), Max.GetValue(X, Y));
 	});
@@ -346,7 +346,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Clamp_FLambdaOperation2D(FLam
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Wrap_FLambdaOperation2D(FLambdaOperation2D Value, FLambdaOperation2D Min, FLambdaOperation2D Max)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Wrap(Value.GetValue(X, Y), Min.GetValue(X, Y), Max.GetValue(X, Y));
 	});
@@ -356,7 +356,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Wrap_FLambdaOperation2D(FLamb
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Lerp_FLambdaOperation2D(FLambdaOperation2D A, FLambdaOperation2D B, FLambdaOperation2D Alpha)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FMath::Lerp(A.GetValue(X, Y), B.GetValue(X, Y), Alpha.GetValue(X, Y));
 	});
@@ -366,7 +366,7 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::Lerp_FLambdaOperation2D(FLamb
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::NormalizeToRange_FLambdaOperation2D(FLambdaOperation2D Value, FLambdaOperation2D RangeMin, FLambdaOperation2D RangeMax)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return UKismetMathLibrary::NormalizeToRange(Value.GetValue(X, Y), RangeMin.GetValue(X, Y), RangeMax.GetValue(X, Y));
 	});
@@ -376,13 +376,13 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::NormalizeToRange_FLambdaOpera
 FLambdaOperation2D UFastNoiseGraphFunctionLibrary::PingPong_FLambdaOperation2D(FLambdaOperation2D Value, FLambdaOperation2D Amplitude, FLambdaOperation2D Frequency)
 {
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
-		const float InValue = Value.GetValue(X, Y);
-		const float InAmplitude = Amplitude.GetValue(X, Y);
-		const float InFrequency = Frequency.GetValue(X, Y);
+		const double InValue = Value.GetValue(X, Y);
+		const double InAmplitude = Amplitude.GetValue(X, Y);
+		const double InFrequency = Frequency.GetValue(X, Y);
 
-		return InAmplitude - FMath::Abs(FMath::Fmod(FMath::Abs(InValue * InFrequency), (2.f * InAmplitude)) - InAmplitude);
+		return InAmplitude - FMath::Abs(FMath::Fmod(FMath::Abs(InValue * InFrequency), (2.0 * InAmplitude)) - InAmplitude);
 	});
 	return Operation;
 }
@@ -396,19 +396,19 @@ FLambdaOperation2D UFastNoiseGraphFunctionLibrary::ApplyCurve_FLambdaOperation2D
 	}
 	else
 	{
-		FloatCurve.AddKey(0.f, 0.f);
-		FloatCurve.AddKey(1.f, 1.f);
+		FloatCurve.AddKey(0.0, 0.0);
+		FloatCurve.AddKey(1.0, 1.0);
 	}
 
 	FLambdaOperation2D Operation;
-	Operation.BindLambda([=](const float X, const float Y)
+	Operation.BindLambda([=](const double X, const double Y)
 	{
 		return FloatCurve.Eval(Value.GetValue(X, Y));
 	});
 	return Operation;
 }
 
-float UFastNoiseGraphFunctionLibrary::CalculateNoiseOperation2D(FLambdaOperation2D ValueOperation2D, float X, float Y)
+double UFastNoiseGraphFunctionLibrary::CalculateNoiseOperation2D(FLambdaOperation2D ValueOperation2D, double X, double Y)
 {
 	return ValueOperation2D.GetValue(X, Y);
 }
